+++ backend/apps/accounts/db_stats.py
@@ -5,8 +5,7 @@ from apps.iamstudent.forms import form_labels
 from apps.iamstudent.models import AUSBILDUNGS_TYPEN_COLUMNS, EmailToHospital, EmailToSend, Student
 from apps.ineedstudent.models import Hospital
 
-from .newsletter import Newsletter
-from .user import User
+from .modelss import Newsletter, User
 
 
 class RegisterList(list):
