+++ backend/apps/ineedstudent/models.py
@@ -6,7 +6,7 @@ from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-from apps.accounts.user import User
+from apps.accounts.modelss import User
 from apps.mapview.utils import plzs
 
 # Create your models here.
